+++ Triangle.cs
@@ -12,7 +12,7 @@ namespace NewHomeWork
         public Triangle(int lengthParameter1, int lengthParameter2, int lengthParameter3) : base(lengthParameter1, lengthParameter2, lengthParameter3)
         {
         }
-        public static void TrianglePerimeter(int lengthParameter1, int lengthParameter2, int lengthParameter3, out int trianglePerimeter)
+        public static void TrianglePerimeter(int lengthParameter1, int lengthParameter2, int lengthParameter3, out double trianglePerimeter)
         {
             trianglePerimeter = lengthParameter1 + lengthParameter2 + lengthParameter3;
             Console.WriteLine($" периметр треугольника {trianglePerimeter}");
